ensure status is implmented in a single register
The ap_uint::operator[] does a read modify write to set a single bit. This results in the status to be split into two registers status_i (0x18) and status_o (0x20) for write and read respectively. In this case the status_o offset would be the one containing the over/underflow flags. This has caused subtle bugs in PS applications reading the wrong status register and thus never reporting an overflow.

This commit addresses the above pitfall by ensuring the status is updated in a single write, thus causing HLS to generate single register.