instrumenation: add min_latency measurement
The currently reported latency is the latency of the last processed feature map. Due to the free running nature of the input data generation, this latency measurement can be seen as taken at an indeterministic time from the point of view of, for example, an application on the PS reading the latency register. In practice, due to the delay and timing of reading this register, this measurement is typically read when the system has most likely reached stable throughput.

However the latency at the start of the run can be (and typically is) lower than when stable. This is relevant in cases where the real world input data rate is lower than what the FINN IP is capable of, and the latency of individual feature maps is more important than throughput (e.g. in HFT). In these cases a minimum latency measurement is more useful as it more closely resembles real world scenario.